# sys-service/api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Aevons 系统服务接口文档，提供用户、角色、部门、岗位、菜单、字典、语言、配置、通知公告和终端应用等系统管理能力",
        "title": "Aevons System Service API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:10702",
    "basePath": "/api/v1/sys",
    "paths": {
        "/api/v1/sys/conf": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "新增参数配置",
                "parameters": [
                    {
                        "description": "新增参数配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateConfDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/conf/key/{key}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "按配置键获取参数配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "配置键",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/conf/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "查询参数配置列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/conf/page": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "分页查询参数配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/conf/refresh-cache": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "刷新参数配置缓存",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/conf/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "获取参数配置详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "修改参数配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改参数配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateConfDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "参数配置"
                ],
                "summary": "删除参数配置",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/dept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "新增部门",
                "parameters": [
                    {
                        "description": "新增部门",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDeptDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/dept/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "查询部门树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/sys/dept/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "获取部门详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "修改部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改部门",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateDeptDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "部门管理"
                ],
                "summary": "删除部门",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateConfDTO": {
            "type": "object",
            "required": [
                "confKey",
                "confValue",
                "name"
            ],
            "properties": {
                "confKey": {
                    "description": "配置KEY",
                    "type": "string",
                    "maxLength": 64
                },
                "confValue": {
                    "description": "配置值",
                    "type": "string"
                },
                "isSecret": {
                    "description": "是否加密/脱敏（0否 1是）",
                    "type": "integer"
                },
                "isSys": {
                    "description": "是否系统内置（0否 1是）",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 255
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 500
                },
                "scope": {
                    "description": "范围（0公开配置，1登录配置，2后台服务配置）",
                    "type": "integer"
                }
            }
        },
        "dto.CreateDeptDTO": {
            "type": "object",
            "required": [
                "deptName",
                "deptType",
                "parentId",
                "sort",
                "status"
            ],
            "properties": {
                "deptName": {
                    "description": "部门名称",
                    "type": "string",
                    "maxLength": 30
                },
                "deptType": {
                    "description": "类型（1机构，2部门）",
                    "type": "integer"
                },
                "parentId": {
                    "description": "父级编号",
                    "type": "string",
                    "example": "0"
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateConfDTO": {
            "type": "object",
            "required": [
                "confKey",
                "confValue",
                "id",
                "name"
            ],
            "properties": {
                "confKey": {
                    "description": "配置KEY",
                    "type": "string",
                    "maxLength": 64
                },
                "confValue": {
                    "description": "配置值",
                    "type": "string"
                },
                "id": {
                    "description": "编号",
                    "type": "integer"
                },
                "isSecret": {
                    "description": "是否加密/脱敏（0否 1是）",
                    "type": "integer"
                },
                "isSys": {
                    "description": "是否系统内置（0否 1是）",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 255
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 500
                },
                "scope": {
                    "description": "范围（0公开配置，1登录配置，2后台服务配置）",
                    "type": "integer"
                }
            }
        },
        "dto.UpdateDeptDTO": {
            "type": "object",
            "required": [
                "deptName",
                "deptType",
                "id",
                "parentId",
                "sort",
                "status"
            ],
            "properties": {
                "deptName": {
                    "description": "部门名称",
                    "type": "string",
                    "maxLength": 30
                },
                "deptType": {
                    "description": "类型（1机构，2部门）",
                    "type": "integer"
                },
                "id": {
                    "description": "部门编号",
                    "type": "string",
                    "example": "0"
                },
                "parentId": {
                    "description": "父级编号",
                    "type": "string",
                    "example": "0"
                },
                "remark": {
                    "description": "备注",
                    "type": "string",
                    "maxLength": 255
                },
                "sort": {
                    "description": "顺序",
                    "type": "integer"
                },
                "status": {
                    "description": "状态（0正常 1停用）",
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer 访问令牌，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}